version final

+++ src/main/java/com/jasperReport/module/order/repository/OrderEntityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderEntityRepository extends JpaRepository<OrderEntity,Integer> {
+public interface OrderEntityRepository extends JpaRepository<OrderEntity,Long> {
 }
